pyproject.toml
[tool.poetry]
name = "TruAPI"
version = "1.0.0"
description = ""
authors = ["McGuire Technology, LLC"]
readme = "README.md"
packages = [{ include = "truapi" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = { extras = ["email"], version = "^2.9.0" }
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
bcrypt = "^4.0.0,<5.0.0"
python-ulid = "^2.7.0"
jinja2 = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.0"
mkdocstrings = { extras = ["python"], version = "^0.26.0" }
mkdocs-swagger-ui-tag = "^0.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "C0111", # missing-docstring
    "C0103", # invalid-name
    "R0903", # too-few-public-methods
]
